wee update to 2010-12-25
git-svn-id: http://grubutils.googlecode.com/svn@44 d95a8657-0e37-07c3-0af6-720afd5d0260

+++ wee/shared.h
@@ -504,7 +504,7 @@ int memcheck (unsigned long long addr, unsigned long long len);
 int grub_putstr (const char *str);
 unsigned int grub_sleep (unsigned int seconds);
 
-int rawread (unsigned long drive, unsigned long long sector, unsigned long long byte_offset, unsigned long byte_len, unsigned long long buf, unsigned long write);
+int rawread (unsigned long drive, unsigned long long sector, unsigned long byte_offset, unsigned long long byte_len, unsigned long long buf, unsigned long write);
 int devread (unsigned long long sector, unsigned long byte_offset, unsigned long long byte_len, unsigned long long buf, unsigned long write);
 
 /* Parse a device string and initialize the global parameters. */
